extensions/requirements.txt
aiohttp==3.7.4.post0; python_version >= "3.6" and python_full_version >= "3.5.3"
async-timeout==3.0.1; python_version >= "3.6" and python_full_version >= "3.5.3"
asynctest==0.13.0; python_version >= "3.5"
atomicwrites==1.4.0; python_version >= "3.6" and python_full_version < "3.0.0" and sys_platform == "win32" or sys_platform == "win32" and python_version >= "3.6" and python_full_version >= "3.4.0"
attrs==20.3.0; python_version >= "3.6" and python_full_version >= "3.5.3" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6")
beautifulsoup4==4.9.3
cachetools==4.2.1; python_version >= "3.5" and python_version < "4.0" and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0")
certifi==2020.12.5; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
chardet==4.0.0; python_version >= "3.6" and python_full_version >= "3.5.3" and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0")
click==7.1.2; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
colorama==0.4.4; python_version >= "3.6" and python_full_version < "3.0.0" and sys_platform == "win32" or sys_platform == "win32" and python_version >= "3.6" and python_full_version >= "3.5.0"
dataclasses-json==0.5.2; python_version >= "3.6"
dataclasses-yaml @ git+https://github.com/nalabelle/py-dataclasses_yaml.git@main ; python_version >= "3.7" and python_version < "4.0"
datafile @ git+https://github.com/nalabelle/py-datafile.git@main ; python_version >= "3.7" and python_version < "4.0"
decorator==4.4.2; python_version >= "2.6" and python_full_version < "3.0.0" or python_full_version >= "3.2.0"
discord.py==1.6.0; python_full_version >= "3.5.3"
file-secrets @ git+https://github.com/nalabelle/py-file_secrets.git@main ; python_version >= "3.7" and python_version < "4.0"
future==0.18.2; python_version >= "2.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
geocoder==1.38.1
giphypop==0.3
google-api-core==1.26.2; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
google-api-python-client==1.12.8; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.3; python_version >= "3.6"
google-auth==1.28.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
googleapis-common-protos==1.53.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
httplib2==0.19.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
idna==2.10; python_version >= "3.6" and python_full_version >= "3.5.3"
importlib-metadata==3.8.1; python_version < "3.8" and python_version >= "3.6" and (python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.8" or python_full_version >= "3.4.0" and python_version >= "3.6" and python_version < "3.8")
iniconfig==1.1.1; python_version >= "3.6"
marshmallow-enum==1.5.1; python_version >= "3.7" and python_version < "4.0"
marshmallow==3.10.0; python_version >= "3.7" and python_version < "4.0"
multidict==5.1.0; python_version >= "3.6" and python_full_version >= "3.5.3"
mypy-extensions==0.4.3; python_version >= "3.6"
oauthlib==3.1.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
packaging==20.9; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
pluggy==0.13.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
protobuf==3.15.6; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
py-rolldice @ git+https://github.com/Fiona1729/py-rolldice.git@master
py==1.10.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
pyasn1-modules==0.2.8; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
pyasn1==0.4.8; python_version >= "3.5" and python_full_version < "3.0.0" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6") or python_version >= "3.5" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6") and python_full_version >= "3.6.0"
pyparsing==2.4.7; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
pyrfc3339==1.1
pytest-asyncio==0.14.0; python_version >= "3.5"
pytest==6.2.2; python_version >= "3.6"
python-dateutil==2.8.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.3.0")
python-forecastio==1.4.0
pytz==2020.5
pyyaml==5.4.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
ratelim==0.1.6
regex==2021.3.17
requests-oauthlib==1.3.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
requests==2.25.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
responses==0.13.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
rsa==4.7.2; python_version >= "3.5" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
six==1.15.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
soupsieve==2.2.1; python_version >= "3.6"
stringcase==1.2.0; python_version >= "3.7" and python_version < "4.0"
toml==0.10.2; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
typing-extensions==3.7.4.3; python_version >= "3.6" and python_full_version >= "3.5.3" and python_version < "3.8"
typing-inspect==0.6.0; python_version >= "3.7" and python_version < "4.0"
uritemplate==3.0.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
urllib3==1.26.4; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "4"
yarl==1.6.3; python_version >= "3.6" and python_full_version >= "3.5.3"
zipp==3.4.1; python_version < "3.8" and python_version >= "3.6"
